fix: extract pwa and api app schematics from traxion workspace
