+++ libs/schematics/src/generators/api-app/files/src/app/app.module.ts__template__
