+++ libs/schematics/src/generators/api-app/files/src/main.ts__template__
