+++ libs/schematics/src/generators/api-app/helpers/index.ts
@@ -1,0 +1,4 @@
+export * from './normalized-options.helper';
+export * from './update-files.helper';
+export * from './update-package-json.helper';
+export * from './update-targets.helper';
