+++ libs/schematics/src/generators/api-app/schema.json
@@ -1,0 +1,31 @@
+{
+  "$id": "ApiApp",
+  "$schema": "http://json-schema.org/schema",
+  "cli": "nx",
+  "properties": {
+    "directory": {
+      "alias": "d",
+      "description": "A directory where the project is placed",
+      "type": "string"
+    },
+    "generatedDir": {
+      "default": "generated",
+      "description": "Configure the default generated directory",
+      "type": "string"
+    },
+    "name": {
+      "$default": {
+        "$source": "argv",
+        "index": 0
+      },
+      "description": "",
+      "type": "string",
+      "x-prompt": "What name would you like to use?"
+    }
+  },
+  "required": [
+    "name"
+  ],
+  "title": "",
+  "type": "object"
+}
