+++ libs/schematics/src/generators/pwa-app/files/proxy.conf.json__template__
