+++ libs/schematics/src/generators/pwa-app/files/src/app/app-routing.module.ts__template__
