+++ libs/schematics/src/generators/pwa-app/files/src/app/app.component.html__template__
