+++ libs/schematics/src/generators/pwa-app/files/src/app/app.component.ts__template__
