+++ libs/schematics/src/generators/pwa-app/files/src/app/app.module.ts__template__
