+++ libs/schematics/src/generators/pwa-app/files/src/app/demo/demo-routing.module.ts__template__
