+++ libs/schematics/src/generators/pwa-app/files/src/app/demo/demo.component.html__template__
