+++ libs/schematics/src/generators/pwa-app/files/src/app/demo/demo.component.ts__template__
