+++ libs/schematics/src/generators/pwa-app/files/src/app/demo/demo.module.ts__template__
