+++ libs/schematics/src/generators/pwa-app/files/src/app/demo/features/login-form.component.less__template__
