+++ libs/schematics/src/generators/pwa-app/files/src/app/demo/features/login-form.component.ts__template__
