+++ libs/schematics/src/generators/pwa-app/files/src/app/demo/ui/api-collapse-ui/api-collapse-ui.component.less__template__
