+++ libs/schematics/src/generators/pwa-app/files/src/app/demo/ui/user-infos-ui/user-infos-ui.component.html__template__
