+++ libs/schematics/src/generators/pwa-app/files/src/app/demo/ui/user-infos-ui/user-infos-ui.component.less__template__
