+++ libs/schematics/src/generators/pwa-app/files/src/environments/environment.default.ts__template__
