+++ libs/schematics/src/generators/pwa-app/files/src/environments/environment.prod.ts__template__
