+++ libs/schematics/src/generators/pwa-app/files/src/environments/environment.ts__template__
