+++ libs/schematics/src/generators/pwa-app/files/src/main.ts__template__
