+++ libs/schematics/src/generators/pwa-app/files/tailwind.config.js__template__
