+++ libs/schematics/src/generators/pwa-app/helpers/update-files.helper.ts
@@ -1,0 +1,36 @@
+import { join } from 'path';
+
+import { generateFiles, Tree } from '@nrwl/devkit';
+
+import { NormalizedOptions } from '../schema';
+
+export function updateFiles(tree: Tree, normalizedOptions: NormalizedOptions) {
+  const { appsDir, name } = normalizedOptions;
+  const pwaPath = join(appsDir, name);
+  const appPath = join(pwaPath, 'src', 'app');
+
+  // delete some files
+  tree.delete(join(appPath, 'app.component.html'));
+  tree.delete(join(appPath, 'app.component.less'));
+  tree.delete(join(appPath, 'app.component.spec.ts'));
+  tree.delete(join(appPath, 'app.component.ts'));
+  tree.delete(join(appPath, 'nx-welcome.component.ts'));
+
+  // Generate the pwa files
+  generateFiles(
+    tree,
+    join(__dirname, '..', 'files'),
+    join(appsDir, name),
+    normalizedOptions,
+  );
+
+  // Update polyfill eslint error
+  const polyfill = (
+    tree.read(join(pwaPath, 'src', 'polyfill.ts')) || ''
+  ).toString();
+
+  tree.write(
+    join(pwaPath, 'src', 'polyfill.ts'),
+    polyfill.replaceAll(/\/\*\*\*.*/g, '/**'),
+  );
+}
