+++ libs/schematics/src/generators/pwa-app/schema.json
@@ -1,0 +1,36 @@
+{
+  "$id": "PwaApp",
+  "$schema": "http://json-schema.org/schema",
+  "cli": "nx",
+  "properties": {
+    "apiName": {
+      "default": "api",
+      "description": "Configure the api name (used for proxy)",
+      "type": "string"
+    },
+    "directory": {
+      "alias": "d",
+      "description": "A directory where the project is placed",
+      "type": "string"
+    },
+    "generatedDir": {
+      "default": "generated",
+      "description": "Configure the default generated directory",
+      "type": "string"
+    },
+    "name": {
+      "$default": {
+        "$source": "argv",
+        "index": 0
+      },
+      "description": "",
+      "type": "string",
+      "x-prompt": "What name would you like to use?"
+    }
+  },
+  "required": [
+    "name"
+  ],
+  "title": "",
+  "type": "object"
+}
